Fix multi-output with alternating strategies.
